log(dm): fold retry sampler into log.go
