+++ dm/pkg/log/log_test.go
@@ -32,6 +32,7 @@ import (
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
 	"go.uber.org/zap/zaptest"
+	"go.uber.org/zap/zaptest/observer"
 )
 
 func TestTestLogger(t *testing.T) {
@@ -46,6 +47,35 @@ func TestTestLogger(t *testing.T) {
 	logger.ErrorFilterContextCanceled("the message", zap.Int("number", 123456),
 		zap.Ints("array", []int{7, 8, 9}), ShortError(errors.Annotate(context.Canceled, "extra info")))
 	require.Empty(t, buffer.Stripped())
+}
+
+func TestNewRetrySampleLogger(t *testing.T) {
+	core, observed := observer.New(zap.InfoLevel)
+	base := Logger{zap.New(core).With(zap.String("base", "logger"))}
+	logger := NewRetrySampleLogger(base, zap.String("component", "retry-unit-test"))
+
+	logger.Error("retryable operation failed", zap.Int("retryNum", 1))
+	logger.Error("retryable operation failed", zap.Int("retryNum", 2))
+	logger.Warn("retryable operation failed", zap.Int("retryNum", 3))
+	logger.Error("another retryable operation failed", zap.Int("retryNum", 4))
+
+	entries := observed.All()
+	require.Len(t, entries, 3)
+
+	require.Equal(t, "retryable operation failed", entries[0].Message)
+	require.Equal(t, zap.ErrorLevel, entries[0].Level)
+	require.Equal(t, map[string]any{
+		"base":      "logger",
+		"component": "retry-unit-test",
+		"retryNum":  int64(1),
+		"sampled":   "",
+	}, entries[0].ContextMap())
+
+	require.Equal(t, "retryable operation failed", entries[1].Message)
+	require.Equal(t, zap.WarnLevel, entries[1].Level)
+
+	require.Equal(t, "another retryable operation failed", entries[2].Message)
+	require.Equal(t, zap.ErrorLevel, entries[2].Level)
 }
 
 // makeTestLogger creates a Logger instance which produces JSON logs.
